pyproject.toml
[project]
name = "abhi-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "bs4>=0.0.2",
    "einops>=0.8.1",
    "fastapi>=0.128.0",
    "langchain>=1.1.3",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.0",
    "langchain-google-genai>=4.0.0",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.3",
    "langchain-pinecone>=0.2.13",
    "langchain-postgres>=0.0.16",
    "langchain-tavily>=0.2.14",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-postgres>=3.0.2",
    "langsmith>=0.4.59",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pillow>=12.0.0",
    "psycopg2>=2.9.11",
    "psycopg[binary,pool]>=3.3.2",
    "pymupdf>=1.26.7",
    "pypdf>=6.4.2",
    "sentence-transformers>=5.2.0",
    "structlog>=25.5.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
]
